agents/orchestrator/pyproject.toml
[project]
name = "geospot-orchestrator"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.26",
    "langchain-mcp-adapters>=0.1.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
]

[tool.uv]
dev-dependencies = []